Added an OraclePeclRecipe which builds the oci8 extension from Pecl. This requires Oracle's Instant Client Lite & SDK.  These are not provided by the PR and need to be supplied by the end user.  If these are not supplied, the OraclePeclRecipe will be an noop.
Build instructions are the same with one exception, you need to add an
additonal mount instruction to `docker run`, i.e. `-v
/path/to/instant/client&sdk:/oracle`.  The recipe expects the files to
be at `/oracle` and without those files the recipe won't run.

For PHP 5.5 and 5.6, the recipe pulls in oci8 2.0.x.  For PHP 7, it
pulls in 2.1.x.

The recipe pulls in the following Oracle libraries and bundles them with
the resulting PHP binary (in `php/lib`).

 - libclntshcore.so
 - libclntsh.so
 - libipc1.so
 - libmql1.so
 - libnnz12.so
 - libociicus.so
 - libons.so

Please be aware of this and that distributing these files is likely in
violation of Oracles terms and conditions.

Please also be aware that libociicus.so is a US English specific library
from the Instant Client Lite.  If you need multi-language support, you
may need to use the full Instant Client and include additional
libraries.