Added recipes to build UnixODBC, PHP's odbc extension and PHP's pdo_odbc extension.
It's a requirement to build UnixODBC from source because the version packaged with
Ubuntu Trusty is older and there has been a major version change to the library.
Thus any ODBC driver compiled against the newer version of UnixODBC will not be
compatible with the older version.  This can be removed when Ubuntu packages v2.3.0+
of UnixODBC.

Because we compile UnixODBC manually and this happens after PHP is compiled, we cannot
build the extensions as a part of PHP.  To work around this, we copy the extension
files out of the PHP source and build them like a stand-alone PECL library, after
UnixODBC has been built.  This causes a minor issue with the ODBC extension, which
requires a small patch to its build files.  See the link in the comment for more
details.